+++ MODELS/automobile.h
@@ -11,7 +11,7 @@ public:
      * @param for Veicolo(targa, posizione, km, numeroPosti, ingombro)
      * @param for MotoreCombustione(capacitaSerbatoio, litriSerbatoio, tipoCarburante, cilindrata, emissioni)
      */
-    Automobile(string targa, string posizione, double km, double capacitaSerbatoio, double litriSerbatoio,
+    Automobile(string targa, string posizione, unsigned long km, double capacitaSerbatoio, double litriSerbatoio,
                Carburante tipoCarburante, unsigned short cilindrata, unsigned short emissioni, unsigned short numeroPosti = 2, unsigned short ingombro = 3);
 
 };
