+++ MODELS/motoelettrica.h
@@ -15,7 +15,7 @@ public:
      * @param numeroPosti
      * @param ingombro
      */
-    MotoElettrica(string targa, string posizione, double km, double capacitaBatteria, double capacitaAttuale,
+    MotoElettrica(string targa, string posizione, unsigned long km, double capacitaBatteria, double capacitaAttuale,
                   VelocitaRicarica caricaSupportata, unsigned short potenza, unsigned short numeroMotori, bool inCarica = false, Colonnina colonninaAttuale = nessuna,
                   unsigned short numeroPosti = 1, unsigned short ingombro = 2);
 };
