+++ MODELS/motorecombustione.cpp
@@ -1,7 +1,7 @@
 #include "MODELS/motorecombustione.h"
 
 
-MotoreCombustione::MotoreCombustione(string targa, string posizione, double km, unsigned short numeroPosti, unsigned short ingombro,
+MotoreCombustione::MotoreCombustione(string targa, string posizione, unsigned long km, unsigned short numeroPosti, unsigned short ingombro,
                                      double capacitaSerbatoio,double litriSerbatoio, Carburante tipoCarburante, unsigned short cilindrata,
                                      unsigned short emissioni):
     Veicolo(targa, posizione, km, numeroPosti, ingombro), _capacitaSerbatoio(capacitaSerbatoio), _litriSerbatoio(litriSerbatoio),
